pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-scientist-skills"
version = "0.5.0"
description = "AI Scientist research pipeline with multi-model consensus via claude-octopus"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "AI Scientist Skills Contributors"},
]
keywords = ["ai", "research", "claude-code", "skills", "machine-learning", "scientific-research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "datasets>=2.14",
    "transformers>=4.35",
    "requests>=2.31",
    "backoff>=2.2",
    "pymupdf4llm>=0.0.5",
    "PyMuPDF>=1.23",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "rich>=13.0",
]

[project.optional-dependencies]
modal = ["modal>=0.73"]

[project.urls]
Homepage = "https://github.com/stmailabs/aisci"
Repository = "https://github.com/stmailabs/aisci"
Issues = "https://github.com/stmailabs/aisci/issues"

[tool.setuptools.packages.find]
include = ["tools*"]

[tool.setuptools.package-data]
tools = [
    "templates/**/*",
    "templates/*",
]

[project.scripts]
ai-scientist-verify = "tools.verify_setup:main"
ai-scientist-device = "tools.device_utils:main"
ai-scientist-config = "tools.config:main"
ai-scientist-search = "tools.search:main"
ai-scientist-state = "tools.state_manager:main"
ai-scientist-metrics = "tools.metric_parser:main"
ai-scientist-latex = "tools.latex_compiler:main"
ai-scientist-pdf = "tools.pdf_reader:main"
ai-scientist-budget = "tools.budget_estimator:main"
ai-scientist-dashboard = "tools.dashboard:main"
ai-scientist-modal-run = "tools.modal_runner:main"
